+++ deployer.go
@@ -12,7 +12,9 @@ import (
 	"runtime"
 	"strings"
 
+	v "github.com/hashicorp/go-version"
 	"github.com/progrium/go-basher"
+	"github.com/skratchdot/open-golang/open"
 )
 
 var Version string
@@ -84,6 +86,26 @@ func application(
 	os.Exit(status)
 }
 
+func OpenBrowser(args []string) {
+	err := open.Start(args[0])
+	if err != nil {
+		fatal("Can't open browser: '" + err.Error())
+	}
+}
+
+func VersionCompare(args []string) {
+	v0, err := v.NewVersion(args[0])
+	if err != nil {
+		fatal("Can't parse version string" + err.Error())
+	}
+
+	v1, err := v.NewVersion(args[1])
+	if err != nil {
+		fatal("Can't parse version string" + err.Error())
+	}
+	fmt.Println(v0.Compare(v1))
+}
+
 func Checksum(args []string) {
 	if len(args) < 1 {
 		fatal("No algorithm specified")
@@ -110,8 +132,10 @@ func main() {
 	}
 
 	application(map[string]func([]string){
-		"checksum":    Checksum,
-		"bin-version": BinVersion,
+		"checksum":        Checksum,
+		"bin-version":     BinVersion,
+		"browse":          OpenBrowser,
+		"version-compare": VersionCompare,
 	}, []string{
 		"include/circle.bash",
 		"include/cloudbreak.bash",
@@ -123,6 +147,7 @@ func main() {
 		"include/docker.bash",
 		"include/env.bash",
 		"include/fn.bash",
+		"include/migrate.bash",
 		"include/module.bash",
 	}, Asset, false)